[chore] add example
