+++ examples/CMakeLists.txt
@@ -3,7 +3,14 @@ project(annadb_driver_example VERSION 1.0 LANGUAGES CXX)
 
 find_package(cppzmq REQUIRED)
 
-add_executable(annadb_driver_example main.cpp ../src/connection.hpp ../src/TySON.hpp ../src/utils.hpp ../src/query.hpp ../src/query_comparision.hpp)
+add_executable(annadb_driver_example main.cpp
+        ../src/connection.hpp
+        ../src/TySON.hpp
+        ../src/utils.hpp
+        ../src/query.hpp
+        ../src/query_comparision.hpp
+        includes/insert_example.hpp
+        src/insert_example.cpp )
 target_link_libraries(annadb_driver_example cppzmq)
 
 target_compile_options(annadb_driver_example PRIVATE
